Started on handle_BQ
The biggest issue now is how to handle parsing nested parentheses, but
this is a starting point at least.

Sorry I've been so flaky with this project, I've had a bunch of things
come up recently; future projects should be a lot better from the
perspective of me getting things done.